+++ ApplicationManagement/Runtime/Scripts/IReflectisApplicationManager.cs
@@ -1,0 +1,10 @@
+using Reflectis.SDK.Utilities;
+
+namespace Reflectis.SDK.ApplicationManagement
+{
+    public interface IReflectisApplicationManager : IApplicationManager
+    {
+        static new IReflectisApplicationManager Instance { get; protected set; }
+    }
+
+}