+++ Fade/Runtime/Scripts/DestroyOnFadeInit.cs
@@ -1,0 +1,17 @@
+using Reflectis.SDK.Core;
+using UnityEngine;
+
+namespace Reflectis.SDK.Fade
+{
+    public class DestroyOnFadeInit : MonoBehaviour
+    {
+
+        private void Update()
+        {
+            if (SM.GetSystem<IFadeSystem>() != null)
+            {
+                Destroy(gameObject);
+            }
+        }
+    }
+}