+++ Transitions/Runtime/Scripts/AbstractInterpolatorFloatTransitionProvider.cs
@@ -1,13 +1,11 @@
-using DG.Tweening;
-using System.Collections;
-using System.Collections.Generic;
+using Reflectis.SDK.Utilities;
 using System.Linq;
 using System.Threading.Tasks;
 using UnityEngine;
 
 namespace Reflectis.SDK.Transitions
 {
-    public abstract class AbstractFloatTransitionProvider : AbstractTransitionProvider
+    public abstract class AbstractInterpolatorFloatTransitionProvider : AbstractTransitionProvider
     {
         [Header("Transition parameters")]
         [SerializeField, Tooltip("Destination")]
@@ -16,10 +14,10 @@ namespace Reflectis.SDK.Transitions
         private float duration;
         [SerializeField]
         private AnimationCurve ease;
-
         private float defaultValue = 0f;
 
-        private Tween tween;
+        private Interpolator interpolator;
+
         protected virtual void Awake()
         {
             if (ease.keys.Count() == 0)
@@ -27,34 +25,34 @@ namespace Reflectis.SDK.Transitions
                 ease = AnimationCurve.Linear(0, 0, duration, 1);
             }
             defaultValue = Getter();
+            interpolator = new Interpolator(duration, LerpFunction, GetStartTime, ease);
         }
+
+        private float GetStartTime()
+        {
+            if (destination == defaultValue)
+            {
+                return duration;
+            }
+            return interpolator.InverseEase.Evaluate((Getter() - defaultValue) / (destination - defaultValue)) * duration;
+        }
+
+        private void LerpFunction(float obj)
+        {
+            Setter(defaultValue + (destination - defaultValue) * obj);
+        }
+
         public override async Task EnterTransitionAsync()
         {
             onEnterTransitionStart?.Invoke();
-            if (tween != null)
-            {
-                tween.Kill();
-            }
-            tween = DOTween.To(Getter, Setter, destination, duration).SetEase(ease);
-            while (tween.IsPlaying())
-            {
-                await Task.Yield();
-            }
+            await interpolator.PlayForward();
             OnEnterTransitionFinish?.Invoke();
         }
 
         public override async Task ExitTransitionAsync()
         {
-            if (tween != null)
-            {
-                tween.Kill();
-            }
             OnExitTransitionStart?.Invoke();
-            tween = DOTween.To(Getter, Setter, defaultValue, duration).SetEase(ease);
-            while (tween.IsPlaying())
-            {
-                await Task.Yield();
-            }
+            await interpolator.PlayBackwards();
             onExitTransitionFinish?.Invoke();
         }