+++ Utilities/Runtime/Scripts/IApplicationManager.cs
@@ -1,13 +1,13 @@
+using System.Threading.Tasks;
+
 namespace Reflectis.SDK.Utilities
 {
     public interface IApplicationManager
     {
-        public static IApplicationManager Instance { get; protected set; }
+        static IApplicationManager Instance { get; protected set; }
 
-        public void QuitApplication();
-
-        public void ErasePlayerSessionData();
-
-        public void HideEverything();
+        void QuitApplication();
+        void ErasePlayerSessionData();
+        Task<bool> CheckInternetConnection();
     }
 }